stdlib: Move to own pkg
